+++ test/cmd/txsim/cli_test.go
@@ -6,10 +6,14 @@ import (
 	"testing"
 	"time"
 
+	"github.com/celestiaorg/celestia-app/app"
+	"github.com/celestiaorg/celestia-app/app/encoding"
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
+	"github.com/celestiaorg/celestia-app/test/genesis"
 	"github.com/celestiaorg/celestia-app/test/util/testfactory"
 	"github.com/celestiaorg/celestia-app/test/util/testnode"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 )
 
@@ -54,13 +58,18 @@ func setup(t testing.TB) (keyring.Keyring, string, string) {
 	}
 	t.Helper()
 
+	cdc := encoding.MakeConfig(app.ModuleEncodingRegisters...).Codec
+
 	// set the consensus params to allow for the max square size
 	cparams := testnode.DefaultParams()
 	cparams.Block.MaxBytes = int64(appconsts.DefaultSquareSizeUpperBound*appconsts.DefaultSquareSizeUpperBound) * appconsts.ContinuationSparseShareContentSize
 
 	cfg := testnode.DefaultConfig().
 		WithConsensusParams(cparams).
-		WithFundedAccounts(testfactory.TestAccName)
+		WithFundedAccounts(testfactory.TestAccName).
+		WithGenesisOptions(
+			genesis.FundAccounts(cdc, []sdk.AccAddress{testnode.TestAddress()}, sdk.NewCoin(app.BondDenom, sdk.NewIntFromUint64(1e15))),
+		)
 
 	cctx, rpcAddr, grpcAddr := testnode.NewNetwork(t, cfg)
