+++ test/genesis/genesis_options.go
@@ -9,6 +9,8 @@ import (
 	qgbtypes "github.com/celestiaorg/celestia-app/x/qgb/types"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 	v1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 )
@@ -50,6 +52,40 @@ func SetDataCommitmentWindow(codec codec.Codec, window uint64) GenesisOption {
 		qgbGenState := qgbtypes.DefaultGenesis()
 		qgbGenState.Params.DataCommitmentWindow = window
 		state[qgbtypes.ModuleName] = codec.MustMarshalJSON(qgbGenState)
+		return state
+	}
+}
+
+// FundAccounts adds a set of accounts to the genesis and then sets their balance as provided.
+// This is good in the case where you have a separate keyring you want to test against and not
+// use the one generated by the testnet infra.
+func FundAccounts(codec codec.Codec, addresses []sdk.AccAddress, balance sdk.Coin) GenesisOption {
+	return func(state map[string]json.RawMessage) map[string]json.RawMessage {
+		// set the accounts in the genesis state
+		var authGenState authtypes.GenesisState
+		codec.MustUnmarshalJSON(state[authtypes.ModuleName], &authGenState)
+
+		genAccounts := make([]authtypes.GenesisAccount, len(addresses))
+		genBalances := make([]banktypes.Balance, len(addresses))
+		for idx, addr := range addresses {
+			genAccounts[idx] = authtypes.NewBaseAccount(addr, nil, uint64(idx+len(authGenState.Accounts)), 0)
+			genBalances[idx] = banktypes.Balance{Address: addr.String(), Coins: sdk.NewCoins(balance)}
+		}
+
+		accounts, err := authtypes.PackAccounts(genAccounts)
+		if err != nil {
+			panic(err)
+		}
+
+		authGenState.Accounts = append(authGenState.Accounts, accounts...)
+		state[authtypes.ModuleName] = codec.MustMarshalJSON(&authGenState)
+
+		// set the balances in the genesis state
+		var bankGenState banktypes.GenesisState
+		codec.MustUnmarshalJSON(state[banktypes.ModuleName], &bankGenState)
+
+		bankGenState.Balances = append(bankGenState.Balances, genBalances...)
+		state[banktypes.ModuleName] = codec.MustMarshalJSON(&bankGenState)
 		return state
 	}
 }
